// be/binex/src/modules/mn-wallet/transaction/enum/type-translate.enum.ts
export enum TypeTranslateEnum {
  DEPOSIT = 'DEPOSIT',
  TRANSFER = 'TRANSFER',
  WITHDRAW = 'WITHDRAW',
  PAYMENT = 'PAYMENT',
  RECEIVE = 'RECEIVE',
}
export enum StatusTranslateEnum {
  PENDING = 'PENDING',
  PROCESSING = 'PROCESSING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}

export enum MethodTranslateEnum {
  BANK = 'BANK',
  WALLET = 'WALLET',
}